Fix remaining code duplication in hex and random utils
