+++ main/hex_utils.c
@@ -2,7 +2,7 @@
 #include <string.h>
 #include <stdio.h>
 
-int hex_digit(char c) {
+static int hex_digit(char c) {
     if (c >= '0' && c <= '9') return c - '0';
     if (c >= 'a' && c <= 'f') return c - 'a' + 10;
     if (c >= 'A' && c <= 'F') return c - 'A' + 10;
